src/requirements.txt
Flask==3.1.0
Werkzeug>=2.2.0,<3.0
psycopg2-binary==2.9.5
PyYAML==6.0